Added all Rift Creatures to be able to spawn, these share the item spawn settings for amount spawned.
Added Purge Items, and Creatures action to allow for removal of creatures and Items, activate a Ebony Wand as power 5 and right click it to see the actions.

Fixed an issue that was causing the mod to not load if other action manipulation mods were present.

Fixed some errors i made in the count checks for Item spawning.